Added some unit tests and fixed bugs that were revealed in the algorithm.
getInterval was returning the wrong type.

+++ src/com/marzhillstudios/quizme/data/Card.java
@@ -66,7 +66,7 @@ public class Card {
         this.count = count;
     }
 
-    public float getInterval() {
+    public int getInterval() {
         return interval;
     }
